undo util functions
