+++ angular/projects/spark-angular/src/lib/components/sprk-autocomplete/directives/sprk-autocomplete-input-container/sprk-autocomplete-input-container.directive.spec.ts
@@ -4,44 +4,7 @@ import { SprkAutocompleteInputContainerDirective } from './sprk-autocomplete-inp
 
 @Component({
   selector: 'sprk-test',
-  template: `
-    <div
-      sprkAutocompleteInputContainer
-      analyticsString="test"
-      ariaLabelledby="test-label"
-      idString="test-id"
-    >
-      <input
-        id="autocomplete-input"
-        name="autocomplete-input"
-        type="text"
-        class="sprk-b-TextInput--has-svg-icon sprk-u-Width-100"
-        aria-describedby="autocomplete--error-container"
-        aria-autocomplete="list"
-        sprkInput
-        idString="autocomplete-input1"
-        autocomplete="off"
-        autocapitalize="off"
-        spellcheck="false"
-      />
-    </div>
-
-    <div sprkAutocompleteInputContainer role="custom" ariaHaspopup="custom">
-      <input
-        id="autocomplete-input"
-        name="autocomplete-input"
-        type="text"
-        class="sprk-b-TextInput--has-svg-icon sprk-u-Width-100"
-        aria-describedby="autocomplete--error-container"
-        aria-autocomplete="list"
-        sprkInput
-        idString="autocomplete-input1"
-        autocomplete="off"
-        autocapitalize="off"
-        spellcheck="false"
-      />
-    </div>
-  `,
+  template: ` <div sprkAutocompleteInputContainer></div> `,
 })
 class TestComponent {}
 
@@ -72,45 +35,5 @@ describe('Spark Autocomplete Input Container Directive', () => {
     expect(element.classList.toString()).toEqual(
       'sprk-b-TextInputIconContainer',
     );
-  });
-
-  it('should add a value for data-analytics if analyticsString has a value', () => {
-    expect(element.getAttribute('data-analytics')).toEqual('test');
-  });
-
-  it('should add a value for data-id if idString has a value', () => {
-    expect(element.getAttribute('data-id')).toEqual('test-id');
-  });
-
-  it('should not add a value for data-analytics if analyticsString has no value', () => {
-    expect(element2.getAttribute('data-analytics')).toEqual(null);
-  });
-
-  it('should not add a value for data-id if idString has no value', () => {
-    expect(element2.getAttribute('data-id')).toEqual(null);
-  });
-
-  it('should add role set to combobox', () => {
-    expect(element.getAttribute('role')).toEqual('combobox');
-  });
-
-  it('should add role set to custom value if supplied', () => {
-    expect(element2.getAttribute('role')).toEqual('custom');
-  });
-
-  it('should add aria-labelledby set to value from input', () => {
-    expect(element.getAttribute('aria-labelledby')).toEqual('test-label');
-  });
-
-  it('should not add aria-labelledby if no value is supplied', () => {
-    expect(element2.getAttribute('aria-labelledby')).toEqual(null);
-  });
-
-  it('should add aria-haspopup set to listbox', () => {
-    expect(element.getAttribute('aria-haspopup')).toEqual('listbox');
-  });
-
-  it('should use custom aria-haspopup value if supplied', () => {
-    expect(element2.getAttribute('aria-haspopup')).toEqual('custom');
   });
 });
